Move IntDateTimeToSting function to utilities module
