+++ src/CMakeLists.txt
@@ -3,3 +3,4 @@ add_subdirectory(gsi-ncdiag)
 add_subdirectory(marine)
 add_subdirectory(met-office)
 add_subdirectory(ncep)
+add_subdirectory(utilities)
