+++ src/met-office/odbapi2nc.py
@@ -6,28 +6,7 @@ import os
 import argparse
 from math import exp
 import ioda_conv_ncio as iconv
-
-#This function takes two integers like date=20180415 time=61532 and converts them to a
-#ISO 8601 standard date/time string like "2018-04-15T06:15:32Z"
-def IntDateTimeToString(date, time):
-    #Make sure passed values are int's since we're counting on integer division
-    date = int(date)
-    time = int(time)
-    #Define consts to prevent new int objects being created every time we use these numbers
-    TEN_THOW = 10000
-    HUNDRED = 100
-
-    year = date // TEN_THOW
-    date = date - year * TEN_THOW
-    month = date // HUNDRED
-    day = date - month * HUNDRED
-
-    hour = time // TEN_THOW
-    time = time - hour * TEN_THOW
-    minute = time // HUNDRED
-    second = time - minute * HUNDRED
-
-    return "%d-%02d-%02dT%02d:%02d:%02dZ" % (year, month, day, hour, minute, second)
+import ioda_conv_util
 
 def ConvertRelativeToSpecificHumidity(rh, rh_err, t, p):
     T_KELVIN = 273.15
@@ -173,10 +152,10 @@ row = c.fetchone()
 refDateTimeString = "UnSeT"
 while row is not None:
     row = FetchRow._make(row)
-    anDateTimeString = IntDateTimeToString(row.andate, row.antime)
+    anDateTimeString = ioda_conv_util.IntDateTimeToString(row.andate, row.antime)
     if (refDateTimeString == "UnSeT"):
         refDateTimeString = anDateTimeString
-    obsDateTimeString = IntDateTimeToString(row.date, row.time)
+    obsDateTimeString = ioda_conv_util.IntDateTimeToString(row.date, row.time)
     #Encode the 8 QC bitfields in the ODB API file into a single value for IODA
     qcVal = (row.report_status_active      * 128 +
              row.report_status_passive     *  64 +
