+++ src/utilities/CMakeLists.txt
@@ -1,0 +1,3 @@
+install (PROGRAMS
+  ioda_conv_util.py
+  DESTINATION bin)
